refactor(runtime): red_schema to directory; extract VCS + governance/IAM snapshot builders
Convert runtime/red_schema.rs into a directory module. The dispatcher
core (rewrite_virtual_names, references_system_schema,
is_system_schema_write, red_query, VirtualTableKind + impl,
virtual_table_kind, and the red.* name constants) stays at
red_schema/mod.rs.

Extract two snapshot-builder families as behavior-preserving moves:
- vcs_views.rs: commits/hash_filter, refs (branches+tags), status,
  conflicts, versioned.
- governance_views.rs: registry + history, managed_policies,
  control_events, users, api_keys + fingerprint, control_capabilities,
  policy.actions, policies + policy/iam record + filter/expr renderers.

Signatures unchanged; red_query match arms gain module paths.

Refs #1638
